Create main.cpp
This program is not yet complete. It allows you view and shuffle the list of restaurants, as well as add or remove restaurants from the list. The next step is to ask the user a series of questions that pit one restaurant against one other until all restaurants have been eliminated except one - the winner.